// apps/customer-portal/webapp/src/types/deployments.ts
export type ProjectDeploymentItem = AuditMetadata & {
  id: string;
  name: string;
  description: string | null;
  url: string | null;
  project: IdLabelRef;
  type: IdLabelRef;
  deployedProductCount?: number;
  instanceCount?: number;
}